refactor kernel and add multirow test
