+++ ydb/library/yql/udfs/common/datetime2/test/canondata/result.json
@@ -39,6 +39,11 @@
             "uri": "file://test.test_ImplicitSplit_/results.txt"
         }
     ],
+    "test.test[MultirowBlockTo]": [
+        {
+            "uri": "file://test.test_MultirowBlockTo_/results.txt"
+        }
+    ],
     "test.test[ParseIso8601]": [
         {
             "uri": "file://test.test_ParseIso8601_/results.txt"
