+++ ydb/library/yql/udfs/common/datetime2/test/canondata/test.test_BlockFrom_/results.txt
@@ -160,6 +160,178 @@
                         [
                             "-604800000000"
                         ];
+                        [
+                            "8640000000000000"
+                        ];
+                        #
+                    ]
+                ]
+            }
+        ]
+    }
+]
+[
+    {
+        "Write" = [
+            {
+                "Type" = [
+                    "ListType";
+                    [
+                        "StructType";
+                        [
+                            [
+                                "ts_seconds";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Timestamp"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "ts_msec";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Timestamp"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "ts_usec";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Timestamp"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "ts_empty";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Timestamp"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_days";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_hours";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_minutes";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_seconds";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_msec";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_usec";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_days_overflow";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ];
+                            [
+                                "interval_null";
+                                [
+                                    "OptionalType";
+                                    [
+                                        "DataType";
+                                        "Interval"
+                                    ]
+                                ]
+                            ]
+                        ]
+                    ]
+                ];
+                "Data" = [
+                    [
+                        [
+                            "3875345000000"
+                        ];
+                        [
+                            "3875345000000"
+                        ];
+                        [
+                            "3875345000000"
+                        ];
+                        #;
+                        [
+                            "604800000000"
+                        ];
+                        [
+                            "604800000000"
+                        ];
+                        [
+                            "604800000000"
+                        ];
+                        [
+                            "604800000000"
+                        ];
+                        [
+                            "604800000000"
+                        ];
+                        [
+                            "-604800000000"
+                        ];
                         #;
                         #
                     ]
