+++ ydb/library/yql/udfs/common/datetime2/test/cases/BlockFrom.in.attr
@@ -88,3 +88,93 @@
     }
 }
 
+{
+    "_yql_row_spec" = {
+        "Type" = [
+            "StructType";
+            [
+                [
+                    "fts_seconds";
+                    [
+                        "DataType";
+                        "Uint32"
+                    ]
+                ];
+                [
+                    "fts_msec";
+                    [
+                        "DataType";
+                        "Uint64"
+                    ]
+                ];
+                [
+                    "fts_usec";
+                    [
+                        "DataType";
+                        "Uint64"
+                    ]
+                ];
+                [
+                    "fdays";
+                    [
+                        "DataType";
+                        "Int16"
+                    ]
+                ];
+                [
+                    "fdays_overflow";
+                    [
+                        "DataType";
+                        "Int32"
+                    ]
+                ];
+                [
+                    "fdays_null";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "Int16"
+                        ]
+                    ]
+                ];
+                [
+                    "fhours";
+                    [
+                        "DataType";
+                        "Int32"
+                    ]
+                ];
+                [
+                    "fminutes";
+                    [
+                        "DataType";
+                        "Int32"
+                    ]
+                ];
+                [
+                    "fseconds";
+                    [
+                        "DataType";
+                        "Int32"
+                    ]
+                ];
+                [
+                    "fmsec";
+                    [
+                        "DataType";
+                        "Int64"
+                    ]
+                ];
+                [
+                    "fusec";
+                    [
+                        "DataType";
+                        "Int64"
+                    ]
+                ]
+            ]
+        ]
+    }
+}
+
