+++ ydb/library/yql/udfs/common/datetime2/test/cases/MultirowBlockTo.in.attr
@@ -1,0 +1,98 @@
+{
+    "_yql_row_spec" = {
+        "Type" = [
+            "StructType";
+            [
+                [
+                    "finterval1";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval2";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval3";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval4";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+            ]
+        ]
+    }
+}
+{
+    "_yql_row_spec" = {
+        "Type" = [
+            "StructType";
+            [
+                [
+                    "finterval1";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval2";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval3";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+                [
+                    "finterval4";
+                    [
+                        "OptionalType";
+                        [
+                            "DataType";
+                            "String"
+                        ]
+                    ]
+                ];
+            ]
+        ]
+    }
+}
